# docs/evidence/contracts/iskai-openapi.current.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProfileUpdate": {
        "properties": {
          "avatar_url": {
            "anyOf": [
              {
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "full_name": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "program_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Program Id"
          },
          "specialization_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Specialization Id"
          }
        },
        "title": "ProfileUpdate",
        "type": "object"
      },
      "RoleUpdate": {
        "properties": {
          "role": {
            "pattern": "^(student|faculty|admin|superadmin)$",
            "title": "Role",
            "type": "string"
          },
          "status": {
            "anyOf": [
              {
                "pattern": "^(pending|approved|rejected)$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "required": [
          "role"
        ],
        "title": "RoleUpdate",
        "type": "object"
      },
...
      "UserUpdate": {
        "properties": {
          "department": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department"
          },
          "full_name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Full Name",
            "type": "string"
          },
          "role": {
            "pattern": "^(student|faculty|admin|superadmin)$",
            "title": "Role",
            "type": "string"
          },
          "status": {
            "anyOf": [
              {
                "pattern": "^(pending|approved|rejected)$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "required": [
          "full_name",
          "role"
        ],
        "title": "UserUpdate",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Centralized AI-Powered Thesis Library using Retrieval-Augmented Generation for the College of Computing Studies, Information and Communication Technology, Isabela State University, Echague.",
    "title": "ISU Thesis AI Library API",
    "version": "2.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/analytics/activity": {
      "get": {
        "description": "Return recent audit activity for authorized administrators.",
        "operationId": "recent_activity_analytics_activity_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "description": "The authenticated account is not permitted to perform this action."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Recent Activity",
        "tags": [
          "analytics"
        ]
      }
    },
    "/analytics/logs/system": {
      "get": {
        "description": "Return department-isolated activity logs.",
        "operationId": "get_system_logs_analytics_logs_system_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get System Logs",
        "tags": [
          "analytics"
        ]
      }
    },
    "/analytics/me": {
      "get": {
        "description": "Return the current user's public profile fields.",
        "operationId": "my_profile_analytics_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "The resource does not exist, or is not visible to this account."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "My Profile",
        "tags": [
          "analytics"
        ]
      },
      "put": {
        "description": "Update only the current user's client-editable profile fields.",
        "operationId": "update_my_profile_analytics_me_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "description": "The request was well formed but failed validation."
          },
          "500": {
            "description": "The server could not complete the request."
          },
          "503": {
            "description": "A required backing service is temporarily unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update My Profile",
        "tags": [
          "analytics"
        ]
      }
    },
    "/analytics/overview": {
      "get": {
        "description": "Return full analytics for the admin dashboard.",
        "operationId": "overview_analytics_overview_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "description": "The authenticated account is not permitted to perform this action."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Overview",
        "tags": [
          "analytics"
        ]
      }
    },
...
    "/analytics/users": {
      "get": {
        "description": "List department users for admins or all users for superadmins.",
        "operationId": "list_users_analytics_users_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Users",
        "tags": [
          "analytics"
        ]
      }
    },
    "/analytics/users/{user_id}": {
      "delete": {
        "description": "Delete an authorized target user.",
        "operationId": "delete_user_analytics_users__user_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "The request payload or parameters were rejected."
          },
          "403": {
            "description": "The authenticated account is not permitted to perform this action."
          },
          "404": {
            "description": "The resource does not exist, or is not visible to this account."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "description": "The server could not complete the request."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete User",
        "tags": [
          "analytics"
        ]
      }
    },
    "/analytics/users/{user_id}/details": {
      "put": {
        "description": "Edit an authorized user's name, role, department, and status.",
        "operationId": "update_user_details_analytics_users__user_id__details_put",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "description": "The authenticated account is not permitted to perform this action."
          },
          "404": {
            "description": "The resource does not exist, or is not visible to this account."
          },
          "422": {
            "description": "The request was well formed but failed validation."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update User Details",
        "tags": [
          "analytics"
        ]
      }
    },
    "/analytics/users/{user_id}/role": {
      "put": {
        "description": "Update an authorized target user's role and approval status.",
        "operationId": "update_user_role_analytics_users__user_id__role_put",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "The request payload or parameters were rejected."
          },
          "403": {
            "description": "The authenticated account is not permitted to perform this action."
          },
          "404": {
            "description": "The resource does not exist, or is not visible to this account."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update User Role",
        "tags": [
          "analytics"
        ]
      }
    },
...
  }
}